Introducing pump-stations and pipes
A pumpstation is built by repeatedly pressing the 'b' character.
For each key-press parts of a circle is drawn. When the circle
is complete, the pump-station is ready and will start pumping
fluid into the cell.

An outlet pipe can be constructed by clicking on the side of a cell.